Added Basic 3D Character + Test
The 3D Animated character model has been implemented, though it doesn't have the full functionality, as it's limited to the 8-Bit typed functionality, and not 16-bit onwards; It hasn't been set up to support manual aiming, nor have the advanced abilities such as Dash, Stomp or Power Shots been implemented. This model can and will be replaced later on to include those animations, whether they're imported from blender or manually handled via unity. It's likely that it will be the latter as it's quite evident that the method in which blender exports to unity is flawed and unstable, which often causes undesirable bugs and issues, which aren't particularly encountered when handling animations directly through Unity.