Post Processing Test is Finished.
There are now two prefabs that can be inserted to facilitate the graphics transition. The keys for the ingame shader selecting script are G (8-Bit), H (16-Bit), J (64-Bit) and K (HD). Anyone is welcome to adjust the 3 different shader profiles to suit the needed aesthetic. Alternatively, we can use only ONE post-processing profile and hardcode the properties instead of switching weights between 0 and 1 across the 3 profiles (The script that transitions between them uses lerping). Anyone in SEPA is also welcome to improve upon the the BitTransitionator.